feat(payment-service): publish PaymentCreatedEvent to Kafka on createPayment

+++ payment-service/src/main/java/com/gym/payment_service/events/PaymentCreatedEvent.java
@@ -16,5 +16,5 @@ public class PaymentCreatedEvent {
     private Double finalAmount;
     private LocalDateTime paymentDate;
     private String promotionId;
-    private Double discountPercentage; 
+    private Long discountPercentage;
 }
